Add check in deleteRoom() • Check if authenticated user is not room user

+++ base/views.py
@@ -89,6 +89,9 @@ def updateRoom(request, pk):
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
 
+    if request.user != room.user:
+        return HttpResponse('You are not allowed here!!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
